+++ library/src/commonTest/kotlin/io/github/ismoy/imagepickerkmp/CameraPhotoHandlerTest.kt
@@ -1,0 +1,13 @@
+package io.github.ismoy.imagepickerkmp
+
+// TODO: Re-enable this test once Compose testing is properly configured
+/*
+import kotlin.test.Test
+import kotlin.test.assertEquals
+import kotlin.test.assertNotNull
+import kotlin.test.assertTrue
+
+class CameraPhotoHandlerTest {
+    // Test implementation will be added back once Compose testing is configured
+}
+*/